Improved parseFunction
