+++ uptask/actions/function/fixtures/object-param.ts
