+++ uptask/examples/@build.ts
@@ -1,3 +1,6 @@
+/**
+ * Build a target
+ */
 export function build(target: string) {
   console.log(target)
 }
