+++ uptask/program.ts
@@ -1,6 +1,6 @@
 import { Command } from "commander"
 import { createCommand } from "./actions/command/create.ts"
-import { parseFile } from "./actions/file/parse.ts"
+import { parseFunctions } from "./actions/function/parse.ts"
 import { searchModules } from "./actions/module/search.ts"
 import { helpConfiguration } from "./helpers/program.ts"
 import packageJson from "./package.json" with { type: "json" }
@@ -17,7 +17,7 @@ export function createProgram(pattern?: string): Command {
 
   const modules = searchModules(pattern)
   for (const mod of modules) {
-    for (const func of parseFile(mod.path)) {
+    for (const func of parseFunctions(mod)) {
       program.addCommand(createCommand(func))
     }
   }
